% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{cambiaNA}
\alias{cambiaNA}
\title{Sustituye en un vector los NA por un string}
\usage{
cambiaNA(x, sustituto = "")
}
\arguments{
\item{x}{el vector}

\item{sustituto}{string. Debe de ser longitud 1. Su defecto es "".}
}
\value{
Un vector de misma longitud que x, donde los NA han sido sustituido
  por `sustituto`.
}
\description{
Sustituye en un vector los NA, por el string `sustituto` que debe ser de
longitud 1.
}
\examples{
changeNA(c(NA, "a", "b", NA)
changeNA(c(NA, "a", "b", NA), sustituto = "blanco")
}

