% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centrosUPCT}
\alias{centrosUPCT}
\title{Centros de la UPCT.}
\format{Un dataframe con 9 filas y 3 columnas
\itemize{
\item NombreCentro String. Tal como aparece en el fichero auxiliar.
\item CodigoSIIU String. El código del centro (campo "Unidad") en el fichero
auxiliar de SIIU sobre centros.
\item Acronimo'
}}
\source{
Fichero auxiliar U06414AX0101_02 de SIIU.
}
\usage{
data(centrosUPCT)
}
\description{
Centros de la UPCT.
}
\keyword{datasets}

