% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{diffdataframe}
\alias{diffdataframe}
\title{Performs a diff between two dataframes}
\usage{
diffdataframe(df1, df2, key, file = NULL)
}
\arguments{
\item{df1,}{df2 dataframes. They should have the same columns, even if they
appear in a different order (\code{rbind(df1, df2)} should make sense)}

\item{key}{String vector. A vector of columns names which serve as a unique identifier.}

\item{file}{String. The name of an xlsx file where the differences between
\code{df1} and \code{df2} are written.}
}
\value{
To access the dataframe that contains the differences, the output of
\code{diffdataframe} should be assigned to a variable.
}
\description{
For two dataframes \code{df1} and \code{df2} with a unique identifier
(\code{key}) and the same columns, performs a
diff: it returns (invisibly) a dataframe with the rows in both dataframes which
present some differences, and optionally, write a xlsx file (using the
\code{openxlsx} package) with the three sheets. The first sheet contains the rows
with highlighted  differences, the second sheet contains the rows of
\code{df1} that have no match in \code{df2}, while the third sheet contains
the rows of \code{df2} with no match in \code{df1}.
}
\examples{
library("dplyr")
## consider two dataframes
df1 <- data_frame(x = paste("id", c(1, 1, 2)),
                  year = c("2001", "2002", "2002"),
                  y = letters[1:3])
df2 <- data_frame(x = paste("id", 1:2),
                  year = rep("2002", 2),
                  y = c("a", "c"))
## writes to a xlsx file but does not produce a dataframe:
 diffdataframe(df1, df2, key = c("x", "year"), file = "diff-df1-vs-df2.xlsx")
## write the resulting dataframe to an object for further use:
diff_df1_df2 <- diffdataframe(df1, df2, key = c("x", "year"))
}

