% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{letraNIFE}
\alias{letraNIFE}
\title{Returns the final letter of a DNI or NIE 8 characters string.}
\usage{
letraNIFE(dni)
}
\arguments{
\item{dni}{String vector.}
}
\value{
the final letter of the DNI / NIE. The function is vectorized, so
that it admits a vector of strings and returns a vector. If \code{dni} does
not match the right format for a DNI or NIE, it returns \code{NA}.
}
\description{
The function implements the algorithm described in
\url{http://www.interior.gob.es/web/servicios-al-ciudadano/dni/calculo-del-digito-de-control-del-nif-nie}
to derive the final letter of a 8 characters. The first character in the
string may be either a number or "X", "Y" or "Z". The remaining 7 characters
must be digits. If \code{dni} does not match this format, the function
returns \code{NA}.
}
\examples{
letraNIFE(c("12345678", "1234", "A1234567", "Y1234567", "X1234567"))

}

