% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planes_centros}
\alias{planes_centros}
\title{Planes de la UPCT.}
\format{Un dataframe con 123 filas y 17columnas}
\source{
dwd_estudios.
}
\usage{
data(planes_centros)
}
\description{
Planes de la  UPCT.
}
\keyword{datasets}

