% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{profilingdf}
\alias{profilingdf}
\title{Realiza el profiling de un dataframe}
\usage{
profilingdf(datos)
}
\arguments{
\item{datos}{El dataframe}
}
\value{
Una matriz de strings, que se puede usar en un informe Rmd por
  ejemplo.
}
\description{
Proporciona una matriz que contiene información de profiling sobre el
dataframe. Está sobre todo pensado para dataframes con columnas strings, o
numéricas pero que toman unos pocos valores (típicamente ficheros SIIU). No
proporciona resúmenes numéricos de columnas continuas.
}
\examples{
opada <- data.frame(Nombre = c("Álvaro", "Antonio", "Mari Carmen",
                               "Mathieu"),
                   Sexo = c("Hombre", "Hombre", "Mujer", "Hombre"),
                     Edad = c(25, 25, 23, 25),
                     Nacionalidad = c("720", "720", "720", "250"))
profilingdf(opada)
## para usar en un informe Rmd, salida html:
print(xtable::xtable(profilingdf(opada)), type = "html", include.rownames = FALSE)
}

