% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siiu_codigos_ambito_estudio}
\alias{siiu_codigos_ambito_estudio}
\title{Códigos SIIU de los ámbitos de estudio.}
\format{Un dataframe con 83 filas y 2 columnas
\itemize{
 \item ID. El código ISCED de tres dígitos.
 \item NOMBRE'
}}
\source{
Apartado "Descargas" en SIIU.
}
\usage{
data(siiu_codigos_ambito_estudio)
}
\description{
Códigos ISCED de las ámbitos de estudio.
Última modificación según portal SIIU: 29/09/2014.
}
\examples{
## Para obtener los códigos de 2 dígitos, corresponden a los múltiplos de 10
dplyr::filter(siiu_codigos_ambito_estudio, substr(ID, 3, 3) == "0") \%>\%
dplyr::mutate(ID2 = substr(ID, 1, 2))
}
\keyword{datasets}

