% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siiu_codigos_areas_conocimiento}
\alias{siiu_codigos_areas_conocimiento}
\title{Códigos SIIU de las áreas de conocimiento.}
\format{Un dataframe con 202 filas y 2 columnas
\itemize{
 \item ID. El código usado en los ficheros SIIU
 \item NOMBRE'
}}
\source{
Apartado "Descargas" en SIIU.
}
\usage{
data(siiu_codigos_areas_conocimiento)
}
\description{
Códigos de las áreas de conocimiento según SIIU.
Última modificación según portal SIIU: 19/05/2014.
}
\keyword{datasets}

