% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siiu_codigos_categoria_pdi}
\alias{siiu_codigos_categoria_pdi}
\title{Códigos SIIU de las categorías de PDI.}
\format{Un dataframe con 57 filas y 3 columnas
\itemize{
 \item ID. El código usado en los ficheros SIIU.
 \item NOMBRE
\item TipoCentro String. El tipo de centro en el que presta su servicio el
PDI. Admite tres valores "Centros propios" (todos nuestros centros propios),
"Centros privados" (Turismo) y "Centros privados sin ánimo de lucro" (el CUD)'
}}
\source{
Apartado "Descargas" en SIIU.
}
\usage{
data(siiu_codigos_categoria_pdi)
}
\description{
Códigos SIIU de las categorías de PDI. Se usa en los ficheros U064xxRH0101
(pdi centros propios), U064xxRH0104 (pdi centros privados, i.e Turismo) y U064xxRH0105
(pdi centros privados sin ánimo de lucro, i.e CUD).
Última modificación según portal SIIU: 01/12/2010.
}
\keyword{datasets}

