% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siiu_codigos_ccaa}
\alias{siiu_codigos_ccaa}
\title{Códigos SIIU de las comunidades autónomas}
\format{Un dataframe con 18 filas y 3 columnas
\itemize{
 \item CA_ID_CCAA. El código usado en los ficheros SIIU
 \item CA_DENOM_CCAA. La denominación de la comunidad autónoma.
 \item CA_SIGLAS_CCAA La sigla de la comunidad autónoma (2 letras)'
}}
\source{
Apartado "Descargas" en SIIU.
}
\usage{
data(siiu_codigos_ccaa)
}
\description{
Códigos de las comunidades autónomas según SIIU.
Última modificación según portal SIIU: 05/10/2010.
}
\keyword{datasets}

