% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siiu_codigos_erasmus}
\alias{siiu_codigos_erasmus}
\title{Códigos ERASMUS de las universidades europeas.}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	5065 obs. of  6 variables:
 $ COD_UNIV_DESTINO: chr  "A  XX" "A BADEN01" "A DORNBIR01" "A EISENST01" ...
 $ PAIS            : chr  NA "AUSTRIA" "AUSTRIA" "AUSTRIA" ...
 $ CIUDAD          : chr  NA "Baden" "Dornbirn" "Eisenstadt" ...
 $ NOMBRE          : chr  "Other - Austria" "PAEDAGOGISCHE HOCHSCHULE NIEDEROESTERREICH" "FACHHOCHSCHULE VORARLBERG GMBH" "STIFTUNG PRIVATE PÄDAGOGISCHE HOCHSCHULE BURGENLAND" ...
 $ SIGLAS_PAIS     : chr  "A" "AT" "AT" "AT" ...
 $ APP_REFER_NUMBER: chr  NA "28505-IC-1-2007-1-AT-ERASMUS-EUC-1" "29270-IC-1-2007-1-AT-ERASMUS-EUCX-1" "28577-IC-1-2007-1-AT-ERASMUS-EUCX-1" ...
}}
\source{
Apartado "Descargas" en SIIU.
}
\usage{
data(siiu_codigos_erasmus)
}
\description{
Códigos ERASMUS de las universidades europeas, contiene el código de la
universidad (COD_UNIV_DESTINO), el nombre del páis ("PAIS"), la ciudad
("CIUDAD"), el nombre de la institución ("NOMBRE"), las siglas del páis
("SIGLAS_PAIS") así como la referencia de la solicitud de adhesión
("APP_REFER_NUMBER").
Última modificación según portal SIIU: 02/02/2015.
}
\examples{
dplyr::filter(siiu_codigos_erasmus, CIUDAD == "Cartagena")
}
\keyword{datasets}

