% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siiu_codigos_paises}
\alias{siiu_codigos_paises}
\title{Códigos SIIU de áreas de los paises}
\format{Un dataframe con 220 filas y 2 columnas
\itemize{
 \item ID El código del páis.
 \item NOMBRE'
}}
\source{
Apartado "Descargas" en SIIU.
}
\usage{
data(siiu_codigos_paises)
}
\description{
Códigos de los paises según SIIU. Corresponden con la codificación ISO
31666-1 numérico. 
Última modificación según portal SIIU: 14/03/2014
}
\keyword{datasets}

