% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{tabla2df}
\alias{tabla2df}
\title{Transforma una tabla 2d en un dataframe con el mismo formato}
\usage{
tabla2df(tabla, cabecerafilas = "")
}
\arguments{
\item{tabla}{La tabla que hay que convertir.}

\item{cabecerafilas}{String. El nombre que se quiere asignar}
}
\value{
Un dataframe que se puede usar para escribir en el fichero Excel con
\code{escribirTablaDatos} o \code{writeDataTable}
}
\description{
Para escribir una tabla en openxlsx usando writeDataTable, es necesario
que el objeto sea un dataframe. Esta función permite transformar una
tabla de dimensión 2 en un dataframe para escribirla.
El argumento cabecerafilas permite especificar en el nombre de cabecera
de la columna que contiene las  filas del dataframe
}
\examples{
v1 <- sample(LETTERS[1:4], 50, replace = TRUE)
v2 <- sample(c("Hombre", "Mujer"), 50, replace = TRUE, prob = c(0.25, 0.75))
x <-tabla2df(table(v1, v2), cabecerafilas = "Categoría")
}

