% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siiu_codigos_erasmus}
\alias{siiu_codigos_erasmus}
\title{Códigos ERASMUS de las universidades europeas.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5065 rows and 6 columns.}
\source{
Apartado "Descargas" en SIIU.
}
\usage{
data(siiu_codigos_erasmus)
}
\description{
Códigos ERASMUS de las universidades europeas, contiene el código de la
universidad (COD_UNIV_DESTINO), el nombre del páis ("PAIS"), la ciudad
("CIUDAD"), el nombre de la institución ("NOMBRE"), las siglas del páis
("SIGLAS_PAIS") así como la referencia de la solicitud de adhesión
("APP_REFER_NUMBER").
Última modificación según portal SIIU: 02/02/2015.
}
\examples{
dplyr::filter(siiu_codigos_erasmus, CIUDAD == "Cartagena")
}
\keyword{datasets}

